drm: samsung: add alpha and blend_mode property
Plane alpha and blending mode are added as plane property.

Coverage and pre-multiplied blending mode are supported
and also can be supported combination with plane alpha.

Under the names of "alpha" and "pixel blend mode", plane alpha
and blending mode properties will be provided by default in future
kernel version.

The same names are used to provide the same user interface
in the future.

Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: If6944262887b4ce341ad8f41b50adb33695289c5