drm: samsung: change window disable method.
Newly requested zpos is used when a plane is enabled.
However, when a plane is disabled, connected window should be used
not newly requested zpos.
Because, a user is not responsible for delivering disabled window number
via zpos when disabling a plane.

And, if previously connected window and newly requested zpos(window)
are different for the same plane when enabling a plane,
connected window should be disabled.

Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I21811cc6a6480a13dd8ac397de91001de566ef7d