drm: samsung: add debugging information of compression source.
If corrupted AFBC data is received as DPU_DMA's input, it operates
HW recovery sequence and generates recovery irq.

At this time, it is important information which HW IP encoded the buffer

DPU_EVT_DMA_RECOVERY event stores compression source of a buffer
and this information is also printed when recovery irq occurs.

Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: Ia1331ee5bb5f7b8f2252d8b8a1e469457ea427f9