drm: samsung: DECON controls sysmmu and power domain.
Sysmmu and power domain control are interconnected in the device tree.
That means it can be controlled at the same time.

It is proper to control DPU power domain by DECON. Because DECON starts
to operate firstly than DSIM and DPP and stops to operate lastly.

Of couse, all the devices such as DECON, DSIM and DPP can control
DPU power domain but it's a waste.

So, sysmmus in DPU are also controlled by DECON.

Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I97f74804216950805bcbe16c9020281a6ba9b724