Transaction: WAR for trigger event counter operations
Originally, if trigger event counter is >= current counter, transaction
submission will fail. There is now a new FR where we would like to
submit transaction "at or after" the trigger counter.

This WAR is to allow transaction submission if trigger counter ==
current counter, in which case, the transaction would turn into an
immediate one.

This FR is still under review and this CL could be reverted, but this is
now needed to enable camera preview.

Bug: 149499104
Change-Id: I521ab249fbc1f92b7001d22c86ebbbe83d9b43a3
Signed-off-by: Edmond Chung <edmondchung@google.com>