Fix lwis_dev_power_up_locked error handling
Currently, when LWIS_DEVICE_ENABLE ioctl is called, if an error occurs,
it is correctly propagated back to userspace, but the device stays
partially enabled. This change modifies the enable call to undo
everything it does in the right order in case it encounters the error.

Oh how I wish for RAII

Bug: 153615429
Change-Id: I9f59fa422a5ead20ce5c7731052295daa3b2bc1b
Signed-off-by: Yuriy Romanenko <yromanenko@google.com>