drm: samsung: remove reference to plane config
Plane config is being used for initialization of plane, so this memory
was passed to exynos_plane_init from stack. Holding a reference to this
memory is invalid as the memory goes away as soon as decon_bind goes out
of scope.

Remove the reference to stack memory and replace with sane values.

Bug: 154144790
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: Ibdd75dd5754396afe8820a4a16f47be05963f6ff