drm: samsung: cleanup bts win config
Currently BTS related win config is reset on every atomic commit, this
is unnecessary if some planes haven't changed. And could be problematic
if only a few planes are part of the atomic commit state. Fix by only
updating BTS win_config for planes that are in commit state and clear
unused windows.

Bug: 157778888
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: If4214d5f1015043ed7c36bd9945fe1bd595d9c1a