Interrupt: Fix race condition in accessing enabled events.
It's possible that while we are inside single_event_enable()
removing entries from the list, an interrupt happens. Then there's a
race condition where the interrupt would be accessing the event that we
just removed, causing an illegal pointer exception.

This fix adds a spinlock in the ISR to prevent the race condition.

Bug: 160792446
Signed-off-by: Edmond Chung <edmondchung@google.com>
Change-Id: I4e208890d1ce7f719dc65a768eae855c31b31786