Revert "Revert "Lwis: Fix event trigger logic.""
This reverts commit 63f41c77c14ff8a7b92ff18868ae325592e513d1.

Reason for revert: After more testing, it seems the original change was not related to the error we saw. The original change fixes the kernel panic while running the CTS test with setprop. Reverting the revert to get the original change in.

Bug: 169169020
Change-Id: I727e2941a391003a523c39b87e14ee981d82c664
(cherry picked from commit 91058946c4f6c32148313f9d995efb1414f04036)
Signed-off-by: J. Avila <elavila@google.com>