logbuffer: move logbuffer and gvotables to kernel
logbuffers are used from the GBMS and TCPCI code to record internal
state and debug information that don't need (or cannot) be routed
to the standard syslog.

votables are used to arbitrate the access and manage the state of a
shared resource. Also make PMIC voter compat a compile time option

Bug: 166510851
Bug: 166528086
Bug: 164260419
Test: recompile
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Idd8752bbd4b5fe7a1145cf2892884050926107dd