drm: samsung: split to set enhance path and dqe
All DECONs need to set up dither in enhance path to use the dither.
In particular, DECON0 uses display dither in DQE, so DQE setting is
required additionally. The display dither can be set up in duplicate
because there can be a request for setup in other ways like sysfs.
So, the setting of display dither in DQE can be only set in
exynos_dqe_update() now.

Bug: 163600343
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I96446179f89c2e94c8928c46fc25e526a2df6862