drm: samsung: add colormap to decon window management
Although window management is available and unique per crtc based on
plane_mask. We also have a fallback to colormap whenever there are no
planes attached to decon, in that case we also need to reserve at least
one window for colormap. To accomplish this, keep a bitmask of available
windows that is shared across all decons to keep track of which windows
are being used by each display.

Bug: 161557612
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: I4753ffc67ee3aed0af7724c84cb70d6d6e9dd605