google_charger: battery defender
BatteryDefender TEMP-DEFEND. Prevent the battery from resting
at high voltage/temperature for extended amount of time.

When engaged, discharge the battery to a safe level.

The battery is recharged to full when battery temperature fall
under the safe level, after some time that the device is
disconnected or when the device is discharged to a safe level.

The battery can be recharged to full also after restarting the
device.

Bug: 169609447
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Iff6c8aa337c72ae604bfceb6fb541ebdedb103ec
Signed-off-by: Ken Tsou <kentsou@google.com>
(cherry picked from commit d91bd03ee5bac0439a693bcd831d9bd99ac98ab3)