drm: samsung: fixed to set CGC LUT
There are two SRAMs for setting CGC LUT SFRs. If CGC LUT is written
to SFR after CGC is enabled, it will write to SRAM_A. After shadow
update, LUT SFR will be connected to SRAM_B. It can be toggled every
shadow update if SHD_REG_UP_REQ is set. If a valid CGC LUT is set up
only once, it is written to SRAM_A and there is garbage data in
SRAM_B. If the SRAM continues to toggle afterwards, the garbage data
can be applied to CGC hw. So if the same CGC LUT setting is requested
continuously, it will be set twice.

Bug: 163067048
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I7bd5b291a0648bd3758e84e8e0b5b2dddaf1adc9