drm: samsung: add debugfs nodes for DQE control at runtime
The debugfs nodes are added to control regamma, degamma and CGC feature
of DQE hardware. The LUT of regamma, degamma and CGC can be set and
these can show the current LUT setting at runtime. And also verbose
nodes of histogram, gamma matrix and linear matrix are added to check
the configurations at runtime.

Bug: 152081861
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I7b089314a6bbc4f1a5b3504fc0f031d9ab03cd35