drm: samsung: implement enable/disable_vblank
This patch implements enable_vblank and disable_vblank callback
function. It can enable or disable vblank by user request.

The framestart is also enabled or disabled in these callbacks.
The reason is that originally, vstatus interrupt of dsim should be
used for vblank purpose, but dsim should handle various interrupts
with one interrupt line, so it couldn't be enabled or disabled only
for vblank purpose. So framestart interrupt of decon is used instead.
In command mode, framestart interrupt is used for original purpose.

Bug: 156411983
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I91376fda88f01f7058f6d35811d019ce72c059aa