LWIS: refine power up/down error handle
The original power down implementation return directly
if it met any error. It caused the rest resource that
did not be released. So, it should keep releasing the
rest resource and return error when all done.
The power up error handling is to release the resources
that have been allocated. It can be simplified to call
power down function to handle it. Because the power
down function will check the resource is allocated or
not before it releases it.

Bug: 170373591
Test: Flash light, all sensors probed.
Signed-off-by: Holmes Chou <holmeschou@google.com>
Change-Id: I9a0409266812b310ea35e93481c8177476e8a046