drm: samsung: support ATC feature.
ATC(Ambient light adaptive Tone Control) is supported. This is
purposed to enhance the outdoor visiblility and to make better
contrasted image. If pp_en is high, TDR min and max is determined
by ambient and back light. Conversely, if pp_en is low, TDR min
and max must be set directly by the user. And also the driver can
go into hibernation with ATC set. At this time, the ATC SRAM data
must be backed up and restored later.

Bug: 162558271
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I5e8b34d5947aba825690567f2d55358eb423823e