LWIS: Use a gpios list to handle gpios when power up/down
For some use cases, there are not only one enable pin that need to
handle when power up and power down lwis device.
Current implementation stored the acquired gpios in pre-defined
variables. It works, but also limited the number of gpios that
can be used. Use a list to store the acquried gpios to make it
more flexible.

Bug: 170373591
Bug: 172984132
Test: whitefin / AV2 / AV4 has preview
Signed-off-by: Holmes Chou <holmeschou@google.com>
Change-Id: I4ddd7030876b9cd295bfbd56c26d3d225c884576