drm: samsung: use framestart irq as vblank instead of VT_STATUS
The vblank must be enabled and disabled dynamically in video mode.
This prevents that this interrupt continuously wakes up cpu to
handle it. However, disabling the interrupt of dsim causes a problem
that doesn't handle other interrupts of dsim. So, the framestart
interrupt of decon will replace VT_STATUS interrupt of dsim.

Bug: 156411983
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: Id07c7aa6fececdd5d55308ce8468d51a4c5ee16d