drm: samsung: support window partial update feature
The window partial update feature has been implemented. This
feature allows some interesting area on LCD to be updated when
updating frames. The user can request the partial update area
with the property of crtc. The driver reconfigures the source and
destination coordinates and sends commands to the panel for setting
partial update area when partial update area is requested by user.

Bug: 151111908
Test: function works properly
Change-Id: If849682628b7345c0b80579d677288123372459b
Signed-off-by: Chris Lu <luchris@google.com>
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>