drm: samsung: adjust frame start timeout based on fps
Current timeout of 50ms works well for refresh rates as low as 20,
but if it goes any lower (ex. 10hz) it could end up timing out
frequently as the frame may start to fetch in as long as 100ms.
In order to handle any refresh rate, we should wait for at least
a full frame time based on currently set mode.

Change-Id: I7c62edebcf41c4ea5a0de5444a5e330367b30445
Signed-off-by: Adrian Salido <salidoa@google.com>