drm: samsung: add is_partial property to connector
This allows the connector to decide whether to support the partial
update feature. In addition, constraint of the minimum block size
for supporting partial update is also determined by each panel driver
at runtime. And the user can know whether driver supports this feature
or not through is_partial property of a connector instead of
partial_region property of a crtc. Because partial_region property is
always created.

Bug: 151111908
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
Change-Id: I714fc15c6c4e3ac0776cc30061b2fceaaf13d178