drm: samsung: prevent duplicate crtc disable call
As part of atomic commit tail, there are 2 calls to crtc disable,
this because the call to drm_atomic_helper_commit_modeset_disables
disables connectors/encoders first but we need to first stop crtc.
In order to avoid reimplementing this function, keep track of the
enabled state in order to ignore the duplicate disable call.

Fixes: 92119a0b86b1 ("drm: samsung: turn off decon ahead of dsim")
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: If0a81bbd0ba019c2a7859877b64fb20a54045046