power: reset: remove dead code
as part of pmu_alive protection, we added some backward compatibility
code for situations where an older version of el3 firmware was in use,
which didn't support the priv_reg smc. we are well past that stage now
and devices should have updated el3 firmware. so let's remove the
backward compat which is dead code now.

Bug: 169128860
Signed-off-by: Okash Khawaja <okash@google.com>
Change-Id: I520746f86a98d74a5b3dab3281704b4811fb8768