pca9468_charger: reset increment state on valid
Reset the increment state machine (last increment) when the actual
input current (iin) is in the valid range.

->prev_inc was reset when ta_vol ot ta_cur hits the ta voltage or
current limit, when actual current goes over ->iin_cc (the target),
when the charger hits the IIN_LOOP (input limited) or when the device
hits the VFLT_LOOP state (i.e. enter of CV)..

The device almost never hit the IIN_LOOP (input limited), so the
ta_cur wes likely to be raised to the adapter max causing overheat.

Bug: 186969924
Test: check charge curve
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Id2e2615bf13e5940950b12bc9d8e0a90544b9874