panel: s6e3hc3: prevent turn on LHBM in unexpected mode
Start from EVT1, LHBM requires set `freq set` to 120hz. If current mode
isn't 120hz when enabling LHBM, it causes underrun issue. If `freq set`
is changed to non-120hz when LHBM is on, LHBM may not work normally.
Issue also happens if LHBM is turned on in AoD mode or switch to AoD
mode when LHBM is on.

Bug: 190710537
Test: UDFPS stress test
Signed-off-by: Midas Chien <midaschieh@google.com>
Change-Id: I1838d2824b5456d93967897d19be80d4b1d44b3b