panel: s6e3hc3: specify refresh rate in panel idle mode
We may have different refresh rates in panel idle mode, create
a variable for it and specify the value in the panel file. If
the value is 0, it means the idle mode is not supported or not
active. Also notify the sensor-HAL that the refresh rate is
changed while idle.

Bug: 189277837
Test: refresh rate change in panel idle mode can be notified
Change-Id: Iec3a5d59f1a4e02176c5d1a3228606c0ffd33f0f
Signed-off-by: Chris Lu <luchris@google.com>