max77729_pmic: consistency checks for THM functions
Check for TEN=1 in the FG Config register before reading from
the external channel and issue a warning if the register value
is not correct.

Force TEN=1 while reading the exernal channel but restore the
initial value to allow for different fuel gauge configurations.

WARN when cannot restore the TMUX configuration in pmic or the
fuel gauge Config register after reading the thermistor and
issue a SW reset. Panic when device functionalities are
compromised.

WARNING reading the battery_id (TMUX=2), usb connector temperature
(TMUX=1) or battery temperature (TMUX=0) via this debug path
might interfere with FG model loading and lead to umpredictable
results.

Bug: 191319560
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Ia77fa8c1107149135e78da02cffbe4b420bcdecc