LWIS: Abort periodic IO operations ASAP upon cancellation.
There was an issue where a user can assume that no periodic IO reads
after the cancellation, while in reality the IO could already be in
queue right before the cancellation. After this change, any pending
periodic IO operations will be skipped if the periodic IO is already
canceled. Cancellation may wait for one IO operation if it's already
ongoing during cancellation, or wait for whole operations to be done
if it contains more than one writes to prevent partial write.

Bug: 191842455
Test: CTS, GCA
Signed-off-by: Firman Hadi Prayoga <firmanhp@google.com>
Change-Id: I313df15bda1efa1941f688f534df23761a265c15