misc: cs40l2x: Remove redundant disable_irq in trigger index store
The locking provides synchronisation between cs40l2x_cp_trigger_index
and cs40l2x_irq, so the disable of the IRQ is now redundant. Further
more this disable is an issue, since it is done under the lock. As
disable_irq will wait for the IRQ thread to finish, this can
deadlock with the IRQ handler which also takes the lock.

Bug: 193740406
Test: compiles
Change-Id: I4d403d02140995d86fe3df16d998c0cf53d6ae13
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Signed-off-by: Vince Leung <leungv@google.com>