drm: samsung: dsim: keep device link attached after disable
Currently device link is removed and recreated every time pipeline is
being disabled. This causes the device link to be recreated every time
display turns off/on, moreover we don't currently check whether the
device link creation was successful, and this could lead to hw failures.
Ensure that device link is created and maintained across enable/disable
but only destroy/recreate the link if dsim is reattached to a different
decon.

Bug: 194141566
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: I0613e0750aaa402c083010f80ef219befc63ae71